third_party/WebKit/LayoutTests/http/tests/loading/preload-image-srcset-src-preloaded-reverse-order.html

<html>
<head>
<script src="/js-test-resources/js-test.js"></script>
<script>
    window.targetScaleFactor = 1;
</script>
<script src="../resources/srcset-helper.js"></script>
<script src="http://127.0.0.1:8000/resources/slow-script.pl?delay=100"></script>
</head>
<body>
<script>
    shouldBeFalse("internals.isPreloaded('resources/base-image2.png');");
    shouldBeFalse("internals.isPreloaded('resources/base-image3.png');");
    shouldBeTrue("internals.isPreloaded('resources/preload-test.jpg');");
    shouldBeTrue("internals.isPreloaded('resources/image2.png');");
</script>
<div>This test passes if the fallback image (defined in the src attribute) is preloaded, and the 1x versions (defined in srcset) aren't,
    even when srcset is defined first.</div>
<div>The second image is here to make sure isPreloaded is working properly</div>
<img height="100" width="100" srcset="resources/base-image3.png 3x, resources/base-image2.png 2x" src="resources/preload-test.jpg">
<img src="resources/image2.png">
</body>
</html>
