third_party/WebKit/LayoutTests/http/tests/loading/preload-image-srcset-src-preloaded-reverse-order-expected.txt

main frame - didStartProvisionalLoadForFrame
main frame - didCommitLoadForFrame
main frame - didFinishDocumentLoadForFrame
main frame - didHandleOnloadEventsForFrame
main frame - didFinishLoadForFrame
main frame - didStartProvisionalLoadForFrame
main frame - didCommitLoadForFrame
main frame - didFinishDocumentLoadForFrame
main frame - didHandleOnloadEventsForFrame
main frame - didFinishLoadForFrame
PASS internals.isPreloaded('resources/base-image2.png'); is false
PASS internals.isPreloaded('resources/base-image3.png'); is false
PASS internals.isPreloaded('resources/preload-test.jpg'); is true
PASS internals.isPreloaded('resources/image2.png'); is true
This test passes if the fallback image (defined in the src attribute) is preloaded, and the 1x versions (defined in srcset) aren't, even when srcset is defined first.
The second image is here to make sure isPreloaded is working properly
 
